fixed int comparison in render_pixelated

+++ src/render_pixelated.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   render_pixelated.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: anpollan <anpollan@student.hive.fi>        +#+  +:+       +#+        */
+/*   By: jjaaskel <jjaaskel@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/17 16:29:31 by anpollan          #+#    #+#             */
-/*   Updated: 2026/01/17 16:31:05 by anpollan         ###   ########.fr       */
+/*   Updated: 2026/02/06 12:56:51 by jjaaskel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ void	render_pixelated(t_thread_data *data)
 			data->new_frame_started = false;
 		while (*data->keep_rendering)
 		{
-			if ((unsigned int)*data->current_pixelate_scale == data->pixelate_scale)
+			if (*data->current_pixelate_scale == data->pixelate_scale)
 				break ;
 			if (data->app->go_wait)
 			{
